Sun, Feb 20, 2022, 8:10 PM +00:00

+++ credit/credit.c
@@ -11,7 +11,7 @@ int main(void)
     // Initialise power counter
     int i = 0;
     int sum = 0;
-    int x = 0;
+    long x = 0;
 
     do
     {
